api-python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
firebase-admin==6.2.0
joblib==1.3.2
scikit-learn==1.3.2
numpy==1.26.1
python-dotenv==1.0.0
pandas==2.1.2
pydantic==2.5.2
